Service and Guard created in angular to implement login functionality.
Created a service to http login request
Created a guard to authenticate a user when it access a url.
Added to enviroment a var that contains the url of api service.
Added a dependency from ( https://www.npmjs.com/package/@auth0/angular-jwt )to implement the user authentication.

Links to learn how use service and guard of authentication
https://www.toptal.com/angular/angular-6-jwt-authentication

Link to know what guard to do
https://codingpotions.com/angular-seguridad/

Link to know what are enviroment variables and how use it when you want create
a request service to api.
https://medium.com/beautiful-angular/angular-2-and-environment-variables-59c57ba643be